+++ examples/asr_simple.rs
@@ -1,7 +1,7 @@
 use env_logger;
 use futures::{Future, Stream};
 use log::{error, info};
-use luis_sys::{builder::Builder, events::Flags, Result};
+use luis_sys::{builder::RecognizerConfig, events::Flags, Result};
 use std::env;
 use tokio;
 
@@ -10,36 +10,43 @@ fn main() {
     env::set_var("RUST_LOG", "debug");
     env_logger::init();
 
+    info!("Start ASR test...");
+    recognize_test().unwrap();
+    info!("Stop ASR test...");
+}
+
+fn recognize_test() -> Result {
     let flags = Flags::Recognization
         | Flags::Session
         | Flags::Connection
         | Flags::SpeechDetection
         | Flags::Canceled;
-    let factory = Builder::new()
-        .subscription_key("d5504c34dab74874930d3fe9f2925578")
-        .region("eastasia")
-        .language("zh-CN")
-        .audio_file_path("examples/chinese_test.wav")
-        .flags(flags);
-    info!("Start ASR test...");
-    recognize_once(&factory).map_err(|e| dbg!(e)).unwrap();
-    recognize_stream(&factory).map_err(|e| dbg!(e)).unwrap();
-    recognize_json(&factory).map_err(|e| dbg!(e)).unwrap();
-    recognize_text(&factory).map_err(|e| dbg!(e)).unwrap();
-    info!("Stop ASR test...");
+    let mut factory = RecognizerConfig::from_subscription(
+        "d5504c34dab74874930d3fe9f2925578",
+        "eastasia",
+    )?;
+    factory
+        .set_language("zh-CN")?
+        .set_audio_file_path("examples/chinese_test.wav")?
+        .set_flags(flags)?;
+    recognize_once(&factory).map_err(|e| dbg!(e))?;
+    recognize_stream(&factory).map_err(|e| dbg!(e))?;
+    recognize_json(&factory).map_err(|e| dbg!(e))?;
+    recognize_text(&factory).map_err(|e| dbg!(e))?;
+    Ok(())
 }
 
-fn recognize_once(factory: &Builder) -> Result {
+fn recognize_once(factory: &RecognizerConfig) -> Result {
     info!("Synchronous ASR ");
-    let recognizer = factory.build()?;
+    let recognizer = factory.recognizer()?;
     let result = recognizer.recognize()?;
     info!("done: {}", result);
     Ok(())
 }
 
-fn recognize_stream(factory: &Builder) -> Result {
+fn recognize_stream(factory: &RecognizerConfig) -> Result {
     info!("Asynchronous ASR, streaming Event object");
-    let mut reco = factory.build()?;
+    let mut reco = factory.recognizer()?;
     let promise = reco.start()?.for_each(|msg| {
         info!("result: {:?}", msg.into_result());
         Ok(())
@@ -48,9 +55,9 @@ fn recognize_stream(factory: &Builder) -> Result {
     Ok(())
 }
 
-fn recognize_json(factory: &Builder) -> Result {
+fn recognize_json(factory: &RecognizerConfig) -> Result {
     info!("Asynchronous ASR, get json results");
-    let mut reco = factory.build()?;
+    let mut reco = factory.recognizer()?;
     let promise = reco
         .start()?
         .filter(Flags::Recognized)
@@ -65,9 +72,9 @@ fn recognize_json(factory: &Builder) -> Result {
     Ok(())
 }
 
-fn recognize_text(factory: &Builder) -> Result {
+fn recognize_text(factory: &RecognizerConfig) -> Result {
     info!("Asynchronous ASR, get text only results.");
-    let mut reco = factory.build()?;
+    let mut reco = factory.recognizer()?;
     let promise = reco.start()?;
     let promise = promise
         .text()